requirements.txt
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
cloudinary==1.44.2
dj-database-url==3.1.2
Django==6.0.4
django-cloudinary-storage==0.3.0
gunicorn==25.3.0
idna==3.13
packaging==26.2
pillow==12.2.0
psycopg2-binary==2.9.12
python-dotenv==1.2.2
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.12.0
